pyproject.toml
[tool.poetry]
authors = [""]
description = ""
name = "alerta-contrib"
version = "0.1.0"

[tool.poetry.dependencies]
PyJWT = ">=2.0.0"
alerta = "^8.5.1"
alerta-server = {extras = ["mongodb", "postgres"], version = "^8.7.0"}
kombu = "^5.2.3"
mock = "^4.0.3"
mypy = "^0.931"
pre-commit = "^2.17.0"
pylint = "^2.12.2"
pytest = ">=5.4.3"
pytest-cov = "^3.0.0"
python = "^3.9"
python-dotenv = "^0.19.2"
requests-mock = "^1.9.3"
six = "^1.16.0"
twine = "^3.7.1"
wheel = "^0.37.1"

[tool.poetry.dev-dependencies]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
